ENH: Trees now compiling, added flexible trait interface for MPI. (#30)
* Removed hdf5 from cargo
* Added hyksort
* Create trait interface for trees
* Remove c interface for now
* Add placeholders for parallel tests
* Add hyksort as an MPI feature
