+++ hyksort/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "hyksort"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+mpi = "0.6.*"
+rand = "0.8.*"
+superslice = "1"
