+++ hyksort/src/lib.rs
@@ -1,0 +1,8 @@
+//! Hyksort [1] implementation in Rust, a near direct port of [2].
+//!
+//! **References**
+//!
+//! [1] Sundar, H., Malhotra, D., & Biros, G. (2013, June). Hyksort: a new variant of hypercube quicksort on distributed memory architectures. In Proceedings of the 27th international ACM conference on international conference on supercomputing (pp. 293-302).
+//!
+//! [2] https://github.com/hsundar/usort
+pub mod hyksort;
