+++ traits/src/fmm.rs
@@ -1,4 +1,4 @@
 //! FMM traits
 
+pub mod fmmtree;
 pub mod translation;
-pub mod tree;
