+++ traits/src/lib.rs
@@ -2,4 +2,5 @@
 
 pub mod fmm;
 pub mod general;
+pub mod tree;
 pub mod types;
