+++ tree/examples/balanced.rs
@@ -1,0 +1,3 @@
+fn main() {
+    assert!(true)
+}
