--- tree/src/c_api.rs
@@ -1,5 +1,0 @@
-//! C API headers.
-
-pub mod constants;
-pub mod distributed;
-pub mod types;
