--- tree/src/c_api/types/mod.rs
@@ -1,4 +1,0 @@
-//! C API for Rust Types
-pub mod domain;
-pub mod morton;
-pub mod point;
