+++ tree/src/lib.rs
@@ -18,9 +18,6 @@
 //! [4] Chan, T. "Closest-point problems simplified on the RAM", ACM-SIAM Symposium on Discrete
 //! Algorithms (2002)
 
-pub mod c_api;
 pub mod constants;
-pub mod data;
-pub mod distributed;
-pub mod single_node;
+pub mod implementations;
 pub mod types;
